Abstract bodies and header property declarations.
