+++ lib/raml.rb
@@ -8,8 +8,10 @@ require_relative 'raml/exceptions'
 require_relative 'raml/parser'
 require_relative 'raml/parser/include'
 
+require_relative 'raml/mixin/bodies'
 require_relative 'raml/mixin/documentable'
 require_relative 'raml/mixin/global'
+require_relative 'raml/mixin/headers'
 require_relative 'raml/mixin/merge'
 require_relative 'raml/mixin/parent'
 require_relative 'raml/mixin/validation'
